+++ packages/react-ui/js/core/layout.ts
@@ -17,15 +17,25 @@ import {
     Wrap,
     type Node as YogaNode,
 } from 'yoga-layout/load';
-import {computeTextScaleAndWrap, computeDimensionsFromBoundingBox} from './text-helpers.js';
-import {setPositionCenter, setPositionLeft, setPositionRight} from './position-helpers.js';
-import {computeMeshChildTransforms} from './mesh-transform-helpers.js';
-import {buildRoundedRectangleMeshProps, buildNineSliceProps} from './mesh-helpers.js';
-import {propsEqual} from './props-helpers.js';
+import {
+    computeTextScaleAndWrap,
+    computeDimensionsFromBoundingBox,
+} from '../helpers/text-helpers.js';
+import {
+    setPositionCenter,
+    setPositionLeft,
+    setPositionRight,
+} from '../helpers/position-helpers.js';
+import {computeMeshChildTransforms} from '../helpers/mesh-transform-helpers.js';
+import {
+    buildRoundedRectangleMeshProps,
+    buildNineSliceProps,
+} from '../helpers/mesh-helpers.js';
+import {propsEqual} from '../helpers/props-helpers.js';
 import {roundedRectangle, roundedRectangleOutline} from '../rounded-rectangle-mesh.js';
 import {nineSlice} from '../nine-slice.js';
 import {NodeWrapper, Context} from './core.js';
-import {DEFAULT_FONT_SIZE, TEXT_BASE_SIZE} from './constants.js';
+import {DEFAULT_FONT_SIZE, TEXT_BASE_SIZE} from '../constants.js';
 
 /**
  * Layout and scene-graph application helpers.
