+++ packages/react-ui/test/unit/renderer.test.ts
@@ -1,10 +1,10 @@
 import {describe, it, expect} from 'vitest';
-import {propsEqual} from '../../js/renderer/props-helpers';
+import {propsEqual} from '../../js/helpers/props-helpers';
 import {
     setPositionCenter,
     setPositionLeft,
     setPositionRight,
-} from '../../js/renderer/position-helpers';
+} from '../../js/helpers/position-helpers';
 
 // Minimal fake Object3D that records the last setPositionLocal call
 class FakeObject3D {
